Remove workarounds for VSO-1269037.
"Compiler ICE on reference to pointer in constexpr"

vector ~2048: This is clearer with the workaround as _Temp is used repeatedly, but we can extract it.

vector ~3279: This is clearer with the workaround as _Temp is used repeatedly.

xmemory ~1209: Remove the workaround as _Temp is used only once.

xmemory ~1230: This is significantly clearer with the workaround, so make it permanent. (The workaround used `auto& _Pnext`, whereas the original code used pointers-to-pointers.)