Fix formatting of extremely short durations using {}
{} is specified to be equivalent to formatting
using iostreams <<, but for very small durations
we would fail to compile such code because of conversions
to longer durations failing (since we form the hh_mm_ss structure
for all formats, even though in this case it isn't used.

This change should make some of the specifiers that _do_ use that
structure work as well (for example {:%H}, attoseconds).

This problem occurs for any chrono conversion where the source type
can not represent even one period of the target type.