Port MSVC-PR-407767 (#2830)
Allow passing additional compiler flags if the compiler is cl.exe. (This is for passing extra flags to the internal "checked" builds of the compiler, enabling use-after-free memory checking in the compiler itself (at a ~15% throughput cost). These flags don't exist in shipping builds of the compiler, thus there's nothing to do for our GitHub test harness.)