Dramatically shrink the `__KERNEL32Functions` table.
The only direct uses of `__KERNEL32Functions` are in `STOREFUNCTIONPOINTER` and `DYNAMICGETCACHEDFUNCTION`.

The only use of `DYNAMICGETCACHEDFUNCTION` is in `IFDYNAMICGETCACHEDFUNCTION`.

There are only two uses of `IFDYNAMICGETCACHEDFUNCTION`, loading `GetCurrentPackageId` and `GetSystemTimePreciseAsFileTime`.

Therefore, all other table entries are unused.

The comment `// This enum should not change` was introduced by GH 1194, but we were being overly cautious. The enum and table aren't mentioned outside of `stl/src`. Also, while the table was `extern "C"`, it wasn't exported. Therefore, they're purely internal to the separately compiled STL.