Guard `__KERNEL32Functions` machinery with `#if !defined(_ONECORE)`. Fuse declaration and definition.
The definition and initialization of `__KERNEL32Functions` were effectively guarded by `#if !defined(_ONECORE)`, so we should extend that guard to all of the machinery.

When fusing the declaration and definition, we're within `_CRT_BEGIN_C_HEADER` so we don't need `extern "C"`. We also don't need `extern` (the definition already has external linkage, and other TUs don't need this anyways).

For style, initialize with `{}` instead of ` = {nullptr}`. They have the same effect here, but `{}` emphasizes that we want to value-initialize the entire table.