.natvis: Add view for unique_ptr/vector of chars
This PR adds .natvis entries using `std::unique_ptr`/`std::vector` to hold various character types so that these are displayed as null terminated strings instead as generic arrays. This is useful when debugging code that uses one of these types as a character buffer, for example in order to call a Windows API.

NOTE: For std::vector, I considered setting the display string to use `[size()]` or similar as a format specifier in order to better handle the case that the string wasn't null terminated. But this winds up with a worse result if it is null terminated, so I wound up deciding to leave it as-is.