Suppress narrowing warnings in piecewise_constant_distrubution::densities
This function converts a `vector` of `double` to a `vector` of `_Ty`, which is a narrowing conversion when `Ty` is `float`. It's silly to warn about this when the Standard makes it clear that the narrowing will occur, and the user has asked for such explicitly by calling `piecewise_constant_distrubution<float>::densities`.

Fixes #3105.