This mirrors the `src/vctools/crt/github` changes from Andrew Dean's internal MSVC-PR-425243.
Note that that PR must target `prod/be`, so this will result in temporary divergence between microsoft/STL's GitHub repo and the internal `prod/fe` branch (which we usually keep binary-identical at all times). I believe that this is tolerable because the `stl/msbuild` files change so rarely, and even if we do need to make changes (e.g. adding a new source file), there should be no major conflicts.