Fix and test noexcept strengthening.
We need to explicitly mark lambdas as noexcept.

Note that tuple equality and spaceship aren't strengthened.
For _Iterator operator==, I used _Zip_iterator_sentinel_equal
(as the `if constexpr` is an optimization).
For operator<=>, I dropped the strengthening.

I believe that the _Size_closure can be unconditionally noexcept.

In the test, ranges::min isn't strengthened, so we need to
avoid expressing our expected result in terms of that.