Add test coverage for multiple values.
For find_last and find_last_if, this transformation is easy - add additional values with a `.second` component of `1729`, which allows us to verify that we always find the last one.

The find_last_if_not test was following the existing find_if_not test, but that's difficult to adapt to "make sure we find the last one". So I've simply taken the find_last_if test and inverted the predicates - not not cat is cat.