Update bug-report.md for VS 2022, add `/std:c++latest`.
This adds `/std:c++latest` to the example repro. Although it displays a lengthy notice, I believe it's valuable to include here.
(Many bug reports will need it and it's easy for people new to the command line to forget it.)

This also updates the example versions to VS 2022 17.6 Preview 3, since we hadn't updated them since open-sourcing.

Finally, this shows how the Copy Info button can be used, and mentions that "we need only the first two lines".