fixing std::shared_ptr<void> initialization from void*
class `shared_ptr` is optimized to avoid keeping a deleter object for non-primitive types which are detected using `_Can_scalar_delete`.
This class relies on SFINAE to detect if the operator delete is valid for a given pointer type.
It turns out that delete a void* is valid in msvc, though, according to the standard is undefined behaviour (7.6.2.9). 
By specializing for void is possible to prevent direct initialization of `shared_ptr<void>` from a void* and prevent bugs.
Note that `default_deleter<void>` is not valid too.