Add C4868 to _STL_DISABLED_WARNINGS
[warning C4868: compiler may not enforce left-to-right evaluation order in braced initializer list](https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-c4868?view=msvc-170)

Office is seeing instances of C4868 being emitted inside of the STL. Example:
`xutility(5637) : warning C4868: compiler may not enforce left-to-right evaluation order in braced initializer list`

After discussing the issue with Stephan he suggested adding the supression to the STL itself:
```
You can non-intrusively add this to _STL_EXTRA_DISABLED_WARNINGS.
We would also consider a PR to add it to _STL_DISABLED_WARNINGS for
everyone (as we do globally suppress a number of /Wall warnings there).
We basically never have such ordering assumptions in the STL, so this
warning provides no value to us.
```