Simplify locale's category constants.
We need to retain _Locbase and comment it as TRANSITION, ABI because it affects sizeof(locale) due to MSVC's deficient EBCO.
(On x64, sizeof(locale) is 16, but removing _Locbase would shrink it to 8.)

However, there appears to be nothing stopping us from moving the constants into locale, as depicted in the Standard.
I've verified that the DLL's exports are unaffected, and they don't mention _Locbase at all.
(They also don't mention the names of these constants, excluding std::collate, std::ctype, and std::messages which are different.)

I don't believe we need `_PGLOBAL` - it's rarely used in the STL, and never for new code.
crtdefs.h defines it to be `__declspec(process)` for `_M_CEE` (i.e. `/clr` and `/clr:pure`).
https://learn.microsoft.com/en-us/cpp/cpp/process?view=msvc-170 explains
"When compiling with `/clr`, global and static variables are per-process by default and do not need to use `__declspec(process)`."
so this doesn't affect `/clr` at all, which is what matters.
(As for `/clr:pure`, it's vanishingly unlikely to matter - these are *constants*, so if multiple application domains start getting multiple copies, who cares?)

`using category = int;` is mandated by the Standard. We previously said `int` for the constants, but the Standard depicts `category`,
so I'm following that now that the typedef is in scope.

Finally, I'm upgrading `const` to `constexpr` - this is not depicted in the Standard, but I believe it's more consistent with the `static constexpr`
variables we use everywhere these days. Note that \[constexpr.functions\] forbids constexpr-strengthening functions, but it doesn't care about variables.
Using `static constexpr` is what allows the major improvement of dropping the out-of-line definitions - see GH 3381 which made that change to ios_base's base class _Iosb.