Mechanically simplify.
This test originally covered both the compiler and library feature-test macros, and was therefore extremely paranoid about checking the behavior of the preprocessor.
We gave the compiler test coverage to the compiler team a long time ago, so we can simplify the library test coverage now.

This commit is a 100% regex search-and-replace, from:

    #ifndef (__cpp_lib_\w+)
    #error \1 is not defined
    #elif \1 != (\d{6}L)
    #error \1 is not \2
    #else
    STATIC_ASSERT\(\1 == \2\);
    #endif

to:

    #ifdef $1
    STATIC_ASSERT($1 == $2);
    #else
    #error $1 is not defined
    #endif

except with \n between each line of the "from" and "to" (presented here on separate lines for clarity).