chrono cleanups: Use `ios_base` with `_CATCH_IO_`.
The first parameter of `_CATCH_IO_` is used to qualify `badbit`. It's usually `ios_base`, but is allowed to vary for headers that are only dragging in `<iosfwd>`. `chrono` has the full definition, so it should just say `ios_base` here. This was the only affected line.

Found myself during code review.