Product fix for GH 4241 `<format>`: `format("{:a<10}", "")` fails STL internal check `_First < _Last`
When `_First == _Last`, the iterator will compare equal to `default_sentinel_t` and the callsite `_Measure_string_prefix()` will do the right thing, we just need to avoid calling `_Decode_utf()` for an empty range.

This will leave `_Next` as null, which is fine - we only need it when we increment the iterator.

`_Measure_string_prefix_iterator_legacy` already appears to handle empty ranges correctly.