Properly detect MSVC, part 1.
DevCom-10265237 "Parameter pack size mismatch doesn't make the require-expression false (regression)" is an MSVC bug.

This preprocessor logic is guarding an additional constraint that MSVC needs, so I'm updating the `#endif` comment to "^^^ workaround ^^^".

Previously, I believe this code was mentally saying, "we're in concepts code, so we only need to distinguish Clang from MSVC". This is no longer the case. Now we need `!defined(__clang__) && !defined(__EDG__)` to prevent EDG from using a workaround that it doesn't need.