Preprocessor NFC: Guard with `__cpp_lib_is_layout_compatible` and `__cpp_lib_is_pointer_interconvertible`.
This has the same effect, and is still marked with `// TRANSITION, LLVM-48860` as a reminder.

I think this makes the `#endif` comment clearer, as it's separated by a huge distance and contains nested preprocessor logic.

Checking the feature-test macros instead of the compiler also makes it clearer what to do in the future, even without "no workaround" to tell us.