In `_Allocate`, extract an early return for `is_constant_evaluated`.
This transformation has very minor impact. Sometimes, we'll call `is_constant_evaluated` when we would have returned `_Traits::_Allocate(_Bytes)` anyways. (This happens when we don't take the high-`_Align` branch and we aren't x86/x64.) This is an extremely minor cost, and it completely vanishes in release mode. The code simplicity is well worth it.