Remove test coverage for `std::experimental::erase/erase_if`.
This was copy-pasted for `// P1209R0 erase_if(), erase()` below, further enhanced to inspect the return values that were added to the Standard, so we aren't losing any test coverage.

The test coverage for `P0458R2 contains()` used the final contents of the ordered/unordered associative containers, so I'm defining them with copy-list-initialization to match the `const` containers below.