Overhaul `generator`'s `static_assert`s.
Add Standard citations.

"generator's second argument" was bogus; use "generator's selected reference type" to describe italic-monospace-reference. Then say "an actual reference type" for clarity.

"an iterator with \[...\] cannot model indirectly_readable" was confusing because it was describing a bad situation, but the other messages are describing how situations must be good. Rephrase this for clarity.