<random>: Skip reset in normal_distribution::operator() with custom params
`normal_distribution` works by computing a pair of coefficients that
are multiplied with the stddev to compute how far (and in which direction)
the result is from the mean. These coefficients are not reliant on the
actual stddev (or mean), so there is no reason to avoid reusing one
computed during last invocation, even if it was with different params.

AFAIK usage of this overload is rare (personally I think this overload
should've never been standardized), the old code still needlessly penalized
it, in the worst case halving the performance.