`_Max_limit`: Use it in `_Mul_overflow`.
This is `_Integer_like _Int`, which is ok because we handle `_Signed128` and `_Unsigned128`.

The first case is simple: `_Int` is known to be unsigned, and we want the maximum.

The second case requires a bit of thought. `_Int` is known to be signed, and `_UInt` is its unsigned counterpart. This code wants the maximum `_Int`, stored in `_UInt` type. The new function makes this clearer than the old math.