Stop calling no-op `_Cnd_destroy_in_situ()`.
`~condition_variable()` can be defaulted, making it trivially destructible.

`~condition_variable_any()` can also be defaulted. It's destroying a `shared_ptr`, so it's not trivially destructible.

Stop declaring `_Cnd_destroy_in_situ()` in xthreads.h. Note that this doesn't affect the dllexport surface, as we previously ensured that `_CRTIMP2_PURE` always appears on both declarations and definitions.

Comment the definition of `_Cnd_destroy_in_situ()` in cond.cpp as preserved for bincompat.