Simplify `__std_atomic_has_cmpxchg16b`: cmpxchg16b is always available for 64-bit Win8.1 / Server 2012 R2.
https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-isprocessorfeaturepresent
> PF_COMPARE_EXCHANGE128
> The atomic compare and exchange 128-bit operation (cmpxchg16b) is available.

https://support.microsoft.com/en-us/windows/system-requirements-2f327e5a-2bae-4011-8848-58180a4353a7
> To install a 64-bit OS on a 64-bit PC, your processor needs to support CMPXCHG16b, PrefetchW, and LAHF/SAHF

As mentioned by
https://en.wikipedia.org/wiki/Windows_8
and confirmed by
https://learn.microsoft.com/en-us/lifecycle/products/windows-8
https://learn.microsoft.com/en-us/lifecycle/products/windows-embedded-8-standard
support for Win8.0 client ended in Jan 2016, with the exception of Windows Embedded 8 Standard which ended in Jul 2023.

The one we have to worry about is Server 2012 classic, which is receiving extended security updates through Oct 2026:
https://learn.microsoft.com/en-us/lifecycle/products/windows-server-2012
But only C++20 `atomic_ref<16 bytes>` is physically affected, and only on potato chips, and only for x64. (32-bit servers would be unaffected.) The potential impact is nil.

Server 2012 classic was released in Sep 2012 and was the latest version for only one year, before Server 2012 R2 was released in Oct 2013.

cmpxchg16b support appeared even earlier: in Intel Nehalem released in Nov 2008, and AMD Bulldozer released in Oct 2011.