`__std_atomic_compare_exchange_128` and `__std_atomic_has_cmpxchg16b` should `abort` for 32-bit.
We can do this because they were declared in `<atomic>` for 64-bit only;
there were never any 32-bit callers. (They have to remain dllexported for 32-bit, but that's all.)

Include `<cstdlib>` for `abort` (which we were already calling in this file).

We can drop `__std_atomic_compare_exchange_128_fallback` without ABI impact because it was defined in an unnamed namespace.