Win8 `GetSystemTimePreciseAsFileTime` is always available.
`<__msvc_chrono.hpp>`: Update comment as `system_clock` is now always Precise.

awint.hpp: We don't need to macroize the `__crt` name because we'll directly call the WinAPI function.
We also don't need to declare the `__crt` name anymore. The definition is still dllexported because it's consistently marked as `_CRTIMP2`.

winapisupp.cpp: We don't need to define and store the function pointer, which was in an unnamed namespace, so no ABI impact.
Mark the `__crt` function as preserved for bincompat, directly call the WinAPI function.

xtime.cpp: Directly call the WinAPI function.