Rename colors, part 1.
Old color names were removed, so we have to use the new names.

Adjust `type ColorName` accordingly.

'Old Bugs' used '--color-scale-red-7', which was problematic because it didn't respond to light/dark.
Replace it with the "attention" color, since it doesn't overlap with 'Skipped Libcxx Tests'.
Adjust the Counter in index.html accordingly.

Change the Counters in index.html for Pull Requests and Combined Age to use normal classes from Primer CSS.
This results in some change for dark mode, but it's still sufficiently clear.