Drop "we use SAL annotations" comment.
This was added by GH 2671 on 2022-05-01, then `_Printf_format_string_` was removed by GH 3919 on 2023-08-03.

We can keep using `/analyze`, though. It adds ~4.5 seconds on my machine, but it's not part of the dev inner loop, and Code Format Validation isn't on the critical path (the Early Builds take longer).