Privatize `pair`'s internal non-Standard constructor
It's arguably conforming to leave the internal constructor public,
because it must be called with 4 arguments, and no Standard
constructor would be selected in such a call. However, it's possibly
better to prevent accident calls to implementation details.

Also move the deleted `operator=` overload to the private section for
consistency, which should affects nothing.