Add `.mailmap` file
Git will use this file to determine a canonical name and email address for each committer and/or author email address to display in the output of `git log` and `git blame`. This is useful when people change names or emails or if someone wants to map their different accounts to a single canonical set of info.

See https://git-scm.com/docs/gitmailmap for details.