Hide some non-Standard functions in `locale::id`
The non-Standard constructor and conversion function are likely to
change the behavior of overload resolution, so perhaps they should
be eliminated. But they're currently appearing in the DLL interface and
thus can't be removed until vNext.

This PR attempts to make them only exist when building the STL. The
newly added Standard default constructor and `_Get_index` function are
templates to avoid alternating DLL interface.
(unsure whether this is sufficient or necessary)