Improve sequence container test coverage.
This exercises both adjacent and non-adjacent removals,
inspects the container's entire contents again, and is somewhat
clearer about distinguishing element values from the number of
elements removed. (In the `std::erase` case, the container no longer
contains odd elements, so removing 5 elements is clear.)

This is, of course, the first digits of pi.