Improve debug messages: std::expected.
This splits "expected stores an error, not a value" into:
"operator->() called on a std::expected that stores an error, not a value"
"operator*() called on a std::expected that stores an error, not a value"

And changes "expected stores a value, not an error" into:
"error() called on a std::expected that stores a value, not an error"

This mentions which member function is being called. Because "expected" doesn't really sound like a noun, I've added "std::" qualification and made the phrases less terse.