Enhancement: Remove IDL=2 null asserts from auto_ptr deref/arrow.
It's silly to assert that _Myptr is non-null when we're about to dereference it (or return it for imminent dereferencing).
We purged this category of asserts from the STL aeons ago, but must have missed auto_ptr.
The idea is that null dereferences are just as obvious in the debugger.
Note that this makes the code consistent with unique_ptr and shared_ptr.