Enhancement: Move span::size_bytes() debug checks into span's ctors.
We don't need to push-disable-pop warning C4127 "conditional expression is constant" because we're going to separately handle static vs. dynamic extents.

Add a comment to size_bytes() explaining that we've already checked. (Having constructor checks is sufficient because spans can't grow.)

The checks are moving into `_Span_extent_type`, which has different cases for static vs. dynamic extents. We need to say `sizeof(_Ty)` because we don't have the `element_type` typedef.

In the debug messages, drop qualification from "std::numeric_limits"; the name is obvious, and if we can mention `span` unqualified, surely `numeric_limits` is fine.

For static extents, we can unconditionally `static_assert`, as the throughput cost is trivial. I'm doing this in the constructor, (1) for consistency with dynamic extents below, and (2) because I could imagine pathological code (in STL test suites?) claiming that `span<int, static_cast<size_t>(-2)>` isn't forbidden to simply instantiate as a type - but the moment that an object of such a type is constructed, the constructor arguments must surely be lying, so we're justified in performing a debug check then.

For dynamic extents, `_Span_extent_type` construction performs the runtime check. This is IDL != 0 (as opposed to hardening), so it's okay if we check more frequently than absolutely necessary. (For example, when constructing from a built-in array or a `std::array`, we don't need to worry about overflow.)