Guard std::function CTAD for static call operators with _HAS_CXX23.
MSVC/Clang/EDG now support this feature.

There's a question of whether we should attempt to expand this CTAD downlevel,
to match how static call operators are supported downlevel.
It would probably work, even for CUDA (which doesn't support static call operators,
but also doesn't appear to reject the SFINAE syntax here on sight).
However, std::function CTAD is a "nice to have" feature,
and we aren't obligated to support it downlevel.
Restricting it to C++23 is less risky, and completely avoids any
question of disrupting CUDA (as CUDA 12.4 supports only up to C++20).