Implement tombstones for `basic_string`.
This is special because of the Small String Optimization. In small mode, the string would be usable,
so we need large mode for the tombstone to have any effect.

I'm setting `_Mysize` to zero, to work with hardened preconditions as usual. For the capacity `_Myres`,
I'm using the first value that `basic_string` would use when entering large mode,
as it has a "roundup mask".

`_Large_mode_engaged()` returns `_Myres > _Small_string_capacity`, so `_Small_string_capacity + 1` would be sufficient,
but it seems safer to use the value that `basic_string` would ordinarily use,
as this is less likely to confuse `basic_string` logic now and in the future.
Only the tombstone pointer needs to be impossible.