Build the STL and benchmarks with /nologo.
By default, this makes no difference. But when investigating what the build system is doing with `-DCMAKE_VERBOSE_MAKEFILE=ON`, this suppresses the verbose:

Microsoft (R) C/C++ Optimizing Compiler Version 19.44.34918.1 for x64
Copyright (C) Microsoft Corporation.  All rights reserved.

and extremely verbose:

/std:c++latest is provided as a preview of language features from the latest C++
working draft, and we're eager to hear about bugs and suggestions for improvements.
However, note that these features are provided as-is without support, and subject
to changes or removal as the working draft evolves. See
https://go.microsoft.com/fwlink/?linkid=2045807 for details.