Prepare ranges test for MSVC fix
The compiler recently addressed an [external bug](https://developercommunity.visualstudio.com/t/conditional-operator-gives-wrong-result-/10665330) and along with that is a better system for handling user-authored casts in `decltype` expressions.

The fix above does not completely fix the compiler behavior mentioned in the workaround, it only makes the compiler more consistently return the type determined in cast expressions.