Added a missing case for `__std_is_file_not_found()`.
I noticed some rather peculiar behavior regarding the Win32 API
`GetFileAttributesExW()`. Specifically, I found that upon calling
this function with an invalid path (in my case, a remote URI) in
two separate applications, both would fail as expected. What is
interesting is that - for some inexplicable reason - the error code
returned by `GetLastError()` differed between these two
applications. Specifically, it was always either one of the
following:

  * `ERROR_INVALID_NAME`: The file name, directory name, or volume
    label syntax is incorrect.

  * `ERROR_BAD_PATHNAME`: The specified path is invalid.

When this occurred, I was making use of the throwing
`std::filesystem::exists()` overload. So, you could imagine my
surprise when this function threw an exception on one of the
applications and returned `false` as expected on the other. I have
not spent any real time investigating the cause for this difference
in behavior myself. In any case, though, it is clear that the
`__std_is_file_not_found()` function should have returned `true`
when given an `__std_win_error` value corresponding to either of
the two error codes.

This commit thus modifies the `__std_is_file_not_found()` function
to take into consideration an additional Win32 error code:
`ERROR_BAD_PATHNAME`. Prior to this commit, this value did not have
a unique value assigned to it in the `__std_win_error`
`enum class`. Thus, a new value by the name of
`__std_win_error::_Bad_pathname` was also added in this commit.