Reduce multprec.cpp dependencies to slightly improve throughput.
We can hardcode `64` instead of including `<limits>` for `numeric_limits<unsigned long long>::digits`.

`uint64_t` and `unsigned long long` are synonyms. In fact, this file was already using `unsigned long long` for the definitions, even though the declarations originally said `uint64_t`.

We moved the declarations out of `<random>` so we don't need it anymore. All we need is `<yvals.h>`.