Restructure control flow with a loop instead of a lambda.
This clearly establishes a loop invariant that `buffer_ptr` is non-null, so we don't need to assert it (whereas in the lambda, the assertion was more useful).

Now we only need the type `PSYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX` twice, so we don't need to introduce an alias `buffer_ptr_t`.

This eliminates the (confusing to me) difference between `DWORD buffer_size` outside and `PDWORD buffer_size` inside the lambda.

We can now scope `unsigned int count = 0;` to the successful case where we need it.

In my opinion, this makes the possible function exits clearer, because they're tested in a flat sequence: we succeed, we experience a true failure, or we can't allocate.