`<xloctime>`: Silence CodeQL warning SM03231 about leap year arithmetic
This is cpp/microsoft/public/leap-year/unchecked-after-arithmetic-year-modification,
"Year field changed using an arithmetic operation without checking for leap year".

Only the line with the conditional operator appeared to be emitting this warning,
but to be safe I'm restructuring the code to use if-statements consistently
(the comments make it clearer anyways), and to suppress the CodeQL warning whenever we're performing arithmetic.