Fuse _Discard_block_base into discard_block_engine.
This was very close to the finish line. Instead of switching between
the old discard_block and the newer _Discard_block_base (introduced by GH 4066),
simply make _Nx's type vary.

Introduce _Block_type (commented with TRANSITION, ABI) and static_cast to avoid sign-comparison warnings,
which was previously achieved by switching between the base classes. We know the constants are within range, so this is value-preserving.

As usual, we no longer need to suppress the deprecation warning.