README.md: (major improvement) Extract installer prerequisites.
This overhauls duplicated, disorganized instructions.

It begins by explaining the need to install Insiders, as installing production releases is a common mistake.

It just says "and keep it up to date" without mentioning our exact required compiler version,
as that doesn't affect what you select in the VS Installer.
The version is now enforced by `CMakeLists.txt` with a clear message.

It gives an explicit list of what workload and components to select.
It doesn't bother with verbosely explaining that you can install CMake/Ninja independently.
It lists AddressSanitizer, which we forgot earlier.
It lists Clang immediately, instead of only mentioning it later in the testing sections.
Also it doesn't bother with explaining that you can install Clang independently.

It explains how to install Python with a little more detail.

The duplication between the IDE and Command Prompt steps is now avoided, as the prereqs are exactly the same.

Finally, the testing section can get straight to the point.